Employee Service

+++ src/main/java/com/study/companytracker/service/DepartmentService.java
@@ -1,6 +1,7 @@
 package com.study.companytracker.service;
 
 
+import com.study.companytracker.converter.DepartmentConverter;
 import com.study.companytracker.dto.DepartmentDTO;
 import com.study.companytracker.dto.GenericRestResponse;
 import com.study.companytracker.exception.NotFoundException;
@@ -12,6 +13,8 @@ import com.study.companytracker.util.enums.ResponseMessage;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.util.Optional;
+
 @Service
 @RequiredArgsConstructor
 public class DepartmentService {
@@ -19,9 +22,9 @@ public class DepartmentService {
     private final DepartmentData departmentData;
 
     /**
+     * @return GenericRestResponse with all departments and response details
      * @Author Ahmed Elakeed
      * Service method to get all departments
-     * @return GenericRestResponse with all departments and response details
      */
     public GenericRestResponse<?> getAllDepartment() {
         try {
@@ -44,7 +47,7 @@ public class DepartmentService {
         try {
             Department department = this.departmentData.findDepartmentByName(name).orElseThrow(() -> new NotFoundException("No department found with name : " + name));
             return GenericRestResponse.builder()
-                    .data(department)
+                    .data(DepartmentConverter.toDto(department))
                     .responseMessage(ResponseMessage.SUCCESS)
                     .responseCode(ResponseMessage.SUCCESS.getCode())
                     .build();
@@ -60,16 +63,16 @@ public class DepartmentService {
 
 
     /**
-     * @Author Mo'men Magdy
      * @param departmentId
      * @return GenericRestResponse with a department object and response details
+     * @Author Mo'men Magdy
      */
     public GenericRestResponse<?> getDepartmentById(Long departmentId) {
         try {
             // finding department by id and throw no found exception if not exist
-            Department department = this.departmentData.findDepartmentById(departmentId).orElseThrow(() -> new NotFoundException("No department found with id : " + departmentId));
+            Department department = this.departmentData.findDepartmentById(departmentId).orElseThrow(() -> new NotFoundException("No departments found with id : " + departmentId));
             return GenericRestResponse.builder()
-                    .data(department)
+                    .data(DepartmentConverter.toDto(department))
                     .responseMessage(ResponseMessage.SUCCESS)
                     .responseCode(ResponseMessage.SUCCESS.getCode())
                     .build();
@@ -83,10 +86,11 @@ public class DepartmentService {
         }
     }
 
-    public GenericRestResponse<?> addDepartment(Department department) {
+    public GenericRestResponse<?> saveDepartment(DepartmentDTO departmentDTO) {
+        Department department = ModelMapperUtil.MAPPER().map(departmentDTO, Department.class);
         try {
             return GenericRestResponse.builder()
-                    .data(departmentData.save(department))
+                    .data(this.departmentData.save(department))
                     .responseMessage(ResponseMessage.SUCCESS)
                     .responseCode(ResponseMessage.SUCCESS.getCode())
                     .build();
@@ -102,8 +106,11 @@ public class DepartmentService {
 
     public GenericRestResponse<?> deleteDepartmentById(Long id) {
         try {
-            this.departmentData.findDepartmentById(id).orElseThrow(() -> new NotFoundException("No departments found with id : " + id));
-            this.departmentData.deleteById(id);
+            Optional<Department> department = this.departmentData.findById(id);
+            if (department.isPresent())
+                this.departmentData.deleteById(id);
+            else
+                throw new NotFoundException("No Employees Found with this Id: " + id);
             return GenericRestResponse.builder()
                     .responseMessage(ResponseMessage.SUCCESS)
                     .responseCode(ResponseMessage.SUCCESS.getCode())
